relocating Location into its own module
