+++ open_precision/core/interfaces/sensor_types/global_positioning_system.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from abc import abstractmethod, ABC
 from open_precision.core.interfaces.sensor_types.basic_sensor import BasicSensor
-from open_precision.core.model.position import Location
+from open_precision.core.model.location import Location
 
 
 class GlobalPositioningSystem(BasicSensor, ABC):
